requirements.txt
celery==4.1.0
djangorestframework==3.7.7
Django==1.11.11
mysqlclient==1.3.12
redis==2.10.6
openpyxl==2.5.1